parser/requirements.txt
sanic
PyJWT
sendgrid
pywaves
requests
aiohttp
psycopg2-binary
aiopg
asyncpg
gino
configparser
base58